refactor auth endpoints to auth blueprints
