+++ app/templates/auth/login.html
@@ -23,7 +23,7 @@
     </form>
     <p>
         Forgot Your Password?
-        <a href="{{ url_for('reset_password_request') }}">Click to Reset It</a>
+        <a href="{{ url_for('auth.reset_password_request') }}">Click to Reset It</a>
     </p>
-    <p>New user? <a href="{{ url_for('register')}}">Click to Register</a></p>
+    <p>New user? <a href="{{ url_for('auth.register')}}">Click to Register</a></p>
 {% endblock content %}
